Fix some XMLHttpRequest CORS header issues
* Access-Control-Allow-Headers was being checked on same-origin requests; now it is only checked on CORS requests
* Preflight responses that have an ok status but mismatching Access-Control-Allow-Headers weren't preventing the follow-up request
* Preflight requests were being sent with all headers, instead of only a subset
