+++ lib/jsdom/living/xhr-utils.js
@@ -1,11 +1,24 @@
 "use strict";
 const request = require("request");
 const { EventEmitter } = require("events");
+const Event = require("./generated/Event");
+const ProgressEvent = require("./generated/ProgressEvent");
 const fs = require("fs");
 const { URL } = require("whatwg-url");
 
 const utils = require("../utils");
+const DOMException = require("domexception");
 const xhrSymbols = require("./xmlhttprequest-symbols");
+
+const headerListSeparatorRegexp = /,[ \t]*/;
+const simpleMethods = new Set(["GET", "HEAD", "POST"]);
+const simpleHeaders = new Set(["accept", "accept-language", "content-language", "content-type"]);
+const preflightHeaders = new Set([
+  "access-control-expose-headers",
+  "access-control-allow-headers",
+  "access-control-allow-credentials",
+  "access-control-allow-origin"
+]);
 
 function wrapCookieJarForRequest(cookieJar) {
   const jarWrapper = request.jar();
@@ -44,23 +57,93 @@ function mergeHeaders(lhs, rhs) {
   return rhsParts.concat(lhsParts.filter(p => rhsParts.indexOf(p) < 0)).join(",");
 }
 
-const simpleMethods = new Set(["GET", "HEAD", "POST"]);
-const simpleHeaders = new Set(["accept", "accept-language", "content-language", "content-type"]);
-const preflightHeaders = new Set([
-  "access-control-expose-headers",
-  "access-control-allow-headers",
-  "access-control-allow-credentials",
-  "access-control-allow-origin"
-]);
+function dispatchError(xhr) {
+  const errMessage = xhr[xhrSymbols.properties].error;
+  requestErrorSteps(xhr, "error", new DOMException(errMessage, "NetworkError"));
 
-exports.getRequestHeader = getRequestHeader;
-exports.updateRequestHeader = updateRequestHeader;
-exports.simpleHeaders = simpleHeaders;
-exports.preflightHeaders = preflightHeaders;
+  if (xhr._ownerDocument) {
+    const error = new Error(errMessage);
+    error.type = "XMLHttpRequest";
+
+    xhr._ownerDocument._defaultView._virtualConsole.emit("jsdomError", error);
+  }
+}
+
+function validCORSHeaders(xhr, response, flag, properties, origin) {
+  const acaoStr = response.headers["access-control-allow-origin"];
+  const acao = acaoStr ? acaoStr.trim() : null;
+  if (acao !== "*" && acao !== origin) {
+    properties.error = "Cross origin " + origin + " forbidden";
+    dispatchError(xhr);
+    return false;
+  }
+  const acacStr = response.headers["access-control-allow-credentials"];
+  const acac = acacStr ? acacStr.trim() : null;
+  if (flag.withCredentials && acac !== "true") {
+    properties.error = "Credentials forbidden";
+    dispatchError(xhr);
+    return false;
+  }
+  return true;
+}
+
+function validCORSPreflightHeaders(xhr, response, flag, properties) {
+  if (!validCORSHeaders(xhr, response, flag, properties, properties.origin)) {
+    return false;
+  }
+  const acahStr = response.headers["access-control-allow-headers"];
+  const acah = new Set(acahStr ? acahStr.trim().toLowerCase().split(headerListSeparatorRegexp) : []);
+  const forbiddenHeaders = Object.keys(flag.requestHeaders).filter(header => {
+    const lcHeader = header.toLowerCase();
+    return !simpleHeaders.has(lcHeader) && !acah.has(lcHeader);
+  });
+  if (forbiddenHeaders.length > 0) {
+    properties.error = "Headers " + forbiddenHeaders + " forbidden";
+    dispatchError(xhr);
+    return false;
+  }
+  return true;
+}
+
+function requestErrorSteps(xhr, event, exception) {
+  const properties = xhr[xhrSymbols.properties];
+  const flag = xhr[xhrSymbols.flag];
+
+  properties.readyState = xhr.DONE;
+  properties.send = false;
+
+  setResponseToNetworkError(xhr);
+
+  if (flag.synchronous) {
+    throw exception;
+  }
+
+  xhr.dispatchEvent(Event.create(["readystatechange"]));
+
+  if (!properties.uploadComplete) {
+    properties.uploadComplete = true;
+
+    if (properties.uploadListener) {
+      xhr.upload.dispatchEvent(ProgressEvent.create([event, { loaded: 0, total: 0, lengthComputable: false }]));
+      xhr.upload.dispatchEvent(ProgressEvent.create(["loadend", { loaded: 0, total: 0, lengthComputable: false }]));
+    }
+  }
+
+  xhr.dispatchEvent(ProgressEvent.create([event, { loaded: 0, total: 0, lengthComputable: false }]));
+  xhr.dispatchEvent(ProgressEvent.create(["loadend", { loaded: 0, total: 0, lengthComputable: false }]));
+}
+
+function setResponseToNetworkError(xhr) {
+  const properties = xhr[xhrSymbols.properties];
+  properties.responseCache = properties.responseTextCache = properties.responseXMLCache = null;
+  properties.responseHeaders = {};
+  properties.status = 0;
+  properties.statusText = "";
+}
 
 // return a "request" client object or an event emitter matching the same behaviour for unsupported protocols
 // the callback should be called with a "request" response object or an event emitter matching the same behaviour too
-exports.createClient = function createClient(xhr) {
+function createClient(xhr) {
   const flag = xhr[xhrSymbols.flag];
   const properties = xhr[xhrSymbols.properties];
   const urlObj = new URL(flag.uri);
@@ -212,9 +295,9 @@ exports.createClient = function createClient(xhr) {
 
   const { body } = flag;
   const hasBody = body !== undefined &&
-    body !== null &&
-    body !== "" &&
-    !(ucMethod === "HEAD" || ucMethod === "GET");
+                  body !== null &&
+                  body !== "" &&
+                  !(ucMethod === "HEAD" || ucMethod === "GET");
 
   if (hasBody && !flag.formData) {
     options.body = body;
@@ -253,7 +336,11 @@ exports.createClient = function createClient(xhr) {
 
     const preflightRequestHeaders = [];
     for (const header in requestHeaders) {
-      preflightRequestHeaders[header] = requestHeaders[header];
+      // the only existing request headers the cors spec allows on the preflight request are Origin and Referrer
+      const lcHeader = header.toLowerCase();
+      if (lcHeader === "origin" || lcHeader === "referrer") {
+        preflightRequestHeaders[header] = requestHeaders[header];
+      }
     }
 
     preflightRequestHeaders["Access-Control-Request-Method"] = flag.method;
@@ -281,36 +368,42 @@ exports.createClient = function createClient(xhr) {
     const preflightClient = request(preflightOptions);
 
     preflightClient.on("response", resp => {
-      if (resp.statusCode >= 200 && resp.statusCode <= 299) {
-        const realClient = doRequest();
-        realClient.on("response", res => {
-          for (const header in resp.headers) {
-            if (preflightHeaders.has(header)) {
-              res.headers[header] = Object.prototype.hasOwnProperty.call(res.headers, header) ?
-                                    mergeHeaders(res.headers[header], resp.headers[header]) :
-                                    resp.headers[header];
-            }
-          }
-          client.emit("response", res);
-        });
-        realClient.on("data", chunk => client.emit("data", chunk));
-        realClient.on("end", () => client.emit("end"));
-        realClient.on("abort", () => client.emit("abort"));
-        realClient.on("request", req => {
-          client.headers = realClient.headers;
-          client.emit("request", req);
-        });
-        realClient.on("redirect", () => {
-          client.response = realClient.response;
-          client.emit("redirect");
-        });
-        realClient.on("error", err => client.emit("error", err));
-        client.abort = () => {
-          realClient.abort();
-        };
-      } else {
+      // don't send the real request if the preflight request returned an error
+      if (resp.statusCode < 200 || resp.statusCode > 299) {
         client.emit("error", new Error("Response for preflight has invalid HTTP status code " + resp.statusCode));
+        return;
       }
+      // don't send the real request if we aren't allowed to use the headers
+      if (!validCORSPreflightHeaders(xhr, resp, flag, properties)) {
+        setResponseToNetworkError(xhr);
+        return;
+      }
+      const realClient = doRequest();
+      realClient.on("response", res => {
+        for (const header in resp.headers) {
+          if (preflightHeaders.has(header)) {
+            res.headers[header] = Object.prototype.hasOwnProperty.call(res.headers, header) ?
+                                  mergeHeaders(res.headers[header], resp.headers[header]) :
+                                  resp.headers[header];
+          }
+        }
+        client.emit("response", res);
+      });
+      realClient.on("data", chunk => client.emit("data", chunk));
+      realClient.on("end", () => client.emit("end"));
+      realClient.on("abort", () => client.emit("abort"));
+      realClient.on("request", req => {
+        client.headers = realClient.headers;
+        client.emit("request", req);
+      });
+      realClient.on("redirect", () => {
+        client.response = realClient.response;
+        client.emit("redirect");
+      });
+      realClient.on("error", err => client.emit("error", err));
+      client.abort = () => {
+        realClient.abort();
+      };
     });
 
     preflightClient.on("error", err => client.emit("error", err));
@@ -337,5 +430,16 @@ exports.createClient = function createClient(xhr) {
   }
 
   return client;
-};
+}
 
+exports.headerListSeparatorRegexp = headerListSeparatorRegexp;
+exports.simpleHeaders = simpleHeaders;
+exports.preflightHeaders = preflightHeaders;
+exports.wrapCookieJarForRequest = wrapCookieJarForRequest;
+exports.getRequestHeader = getRequestHeader;
+exports.updateRequestHeader = updateRequestHeader;
+exports.dispatchError = dispatchError;
+exports.validCORSHeaders = validCORSHeaders;
+exports.requestErrorSteps = requestErrorSteps;
+exports.setResponseToNetworkError = setResponseToNetworkError;
+exports.createClient = createClient;
