+++ lib/jsdom/living/xmlhttprequest.js
@@ -26,7 +26,6 @@ const { parseJSONFromBytes } = require("./helpers/json");
 const syncWorkerFile = require.resolve ? require.resolve("./xhr-sync-worker.js") : null;
 
 const tokenRegexp = /^[!#$%&'*+\-.^_`|~0-9A-Za-z]+$/;
-const headerListSeparatorRegexp = /,[ \t]*/;
 const fieldValueRegexp = /^[ \t]*(?:[\x21-\x7E\x80-\xFF](?:[ \t][\x21-\x7E\x80-\xFF])?)*[ \t]*$/;
 
 const forbiddenRequestHeaders = new Set([
@@ -90,8 +89,6 @@ const XMLHttpRequestResponseType = new Set([
   "json",
   "text"
 ]);
-
-const { simpleHeaders } = xhrUtils;
 
 module.exports = function createXMLHttpRequest(window) {
   const { Event, ProgressEvent } = window;
@@ -385,20 +382,20 @@ module.exports = function createXMLHttpRequest(window) {
         // Run the https://xhr.spec.whatwg.org/#handle-errors "fatal" steps.
         properties.readyState = XMLHttpRequest.DONE;
         properties.send = false;
-        setResponseToNetworkError(this);
+        xhrUtils.setResponseToNetworkError(this);
         return;
       }
 
       if ((this.readyState === XMLHttpRequest.OPENED && properties.send) ||
           this.readyState === XMLHttpRequest.HEADERS_RECEIVED ||
           this.readyState === XMLHttpRequest.LOADING) {
-        requestErrorSteps(this, "abort");
+        xhrUtils.requestErrorSteps(this, "abort");
       }
 
       if (this.readyState === XMLHttpRequest.DONE) {
         properties.readyState = XMLHttpRequest.UNSENT;
 
-        setResponseToNetworkError(this);
+        xhrUtils.setResponseToNetworkError(this);
       }
     }
     getAllResponseHeaders() {
@@ -630,7 +627,7 @@ module.exports = function createXMLHttpRequest(window) {
         this[xhrSymbols.properties] = response.properties;
 
         if (response.properties.error) {
-          dispatchError(this);
+          xhrUtils.dispatchError(this);
           throw new DOMException(response.properties.error, "NetworkError");
         } else {
           const { responseBuffer } = this[xhrSymbols.properties];
@@ -664,7 +661,7 @@ module.exports = function createXMLHttpRequest(window) {
         client.on("error", err => {
           client.removeAllListeners();
           properties.error = err;
-          dispatchError(this);
+          xhrUtils.dispatchError(this);
         });
 
         client.on("response", res => receiveResponse(this, res));
@@ -683,12 +680,12 @@ module.exports = function createXMLHttpRequest(window) {
 
           if (flag.origin !== destUrlObj.origin &&
               destUrlObj.protocol !== "data:") {
-            if (!validCORSHeaders(this, response, flag, properties, flag.origin)) {
+            if (!xhrUtils.validCORSHeaders(this, response, flag, properties, flag.origin)) {
               return;
             }
             if (urlObj.username || urlObj.password) {
               properties.error = "Userinfo forbidden in cors redirect";
-              dispatchError(this);
+              xhrUtils.dispatchError(this);
             }
           }
         });
@@ -835,11 +832,11 @@ module.exports = function createXMLHttpRequest(window) {
     const destUrlObj = new URL(response.request.uri.href);
     if (properties.origin !== destUrlObj.origin &&
         destUrlObj.protocol !== "data:") {
-      if (!validCORSHeaders(xhr, response, flag, properties, properties.origin)) {
+      if (!xhrUtils.validCORSHeaders(xhr, response, flag, properties, properties.origin)) {
         return;
       }
       const acehStr = response.headers["access-control-expose-headers"];
-      const aceh = new Set(acehStr ? acehStr.trim().toLowerCase().split(headerListSeparatorRegexp) : []);
+      const aceh = new Set(acehStr ? acehStr.trim().toLowerCase().split(xhrUtils.headerListSeparatorRegexp) : []);
       for (const header in headers) {
         const lcHeader = header.toLowerCase();
         if (!corsSafeResponseHeaders.has(lcHeader) && !aceh.has(lcHeader)) {
@@ -975,127 +972,50 @@ module.exports = function createXMLHttpRequest(window) {
     });
   }
 
-  function dispatchError(xhr) {
-    const errMessage = xhr[xhrSymbols.properties].error;
-    requestErrorSteps(xhr, "error", new DOMException(errMessage, "NetworkError"));
-
-    if (xhr._ownerDocument) {
-      const error = new Error(errMessage);
-      error.type = "XMLHttpRequest";
-
-      xhr._ownerDocument._defaultView._virtualConsole.emit("jsdomError", error);
-    }
-  }
-
-  function validCORSHeaders(xhr, response, flag, properties, origin) {
-    const acaoStr = response.headers["access-control-allow-origin"];
-    const acao = acaoStr ? acaoStr.trim() : null;
-    if (acao !== "*" && acao !== origin) {
-      properties.error = "Cross origin " + origin + " forbidden";
-      dispatchError(xhr);
-      return false;
-    }
-    const acacStr = response.headers["access-control-allow-credentials"];
-    const acac = acacStr ? acacStr.trim() : null;
-    if (flag.withCredentials && acac !== "true") {
-      properties.error = "Credentials forbidden";
-      dispatchError(xhr);
-      return false;
-    }
-    const acahStr = response.headers["access-control-allow-headers"];
-    const acah = new Set(acahStr ? acahStr.trim().toLowerCase().split(headerListSeparatorRegexp) : []);
-    const forbiddenHeaders = Object.keys(flag.requestHeaders).filter(header => {
-      const lcHeader = header.toLowerCase();
-      return !simpleHeaders.has(lcHeader) && !acah.has(lcHeader);
-    });
-    if (forbiddenHeaders.length > 0) {
-      properties.error = "Headers " + forbiddenHeaders + " forbidden";
-      dispatchError(xhr);
-      return false;
-    }
-    return true;
-  }
-
-  function toByteString(value) {
-    value = String(value);
-    if (!/^[\0-\xFF]*$/.test(value)) {
-      throw new TypeError("invalid ByteString");
-    }
-    return value;
-  }
-
-  function finalMIMEType(xhr) {
-    const flag = xhr[xhrSymbols.flag];
-    return parseContentType(flag.overrideMIMEType || getResponseHeader(xhr, "content-type"));
-  }
-
-  function finalCharset(xhr) {
-    const flag = xhr[xhrSymbols.flag];
-    if (flag.overrideCharset) {
-      return flag.overrideCharset;
-    }
-    const parsedContentType = parseContentType(getResponseHeader(xhr, "content-type"));
-    if (parsedContentType) {
-      return whatwgEncoding.labelToName(parsedContentType.get("charset"));
-    }
-    return null;
-  }
-
-  function getResponseHeader(xhr, lcHeader) {
-    const properties = xhr[xhrSymbols.properties];
-    const keys = Object.keys(properties.responseHeaders);
-    let n = keys.length;
-    while (n--) {
-      const key = keys[n];
-      if (key.toLowerCase() === lcHeader) {
-        return properties.responseHeaders[key];
-      }
-    }
-    return null;
-  }
-
-  function normalizeHeaderValue(value) {
-    return value.replace(/^[\x09\x0A\x0D\x20]+/, "").replace(/[\x09\x0A\x0D\x20]+$/, "");
-  }
-
-  function requestErrorSteps(xhr, event, exception) {
-    const properties = xhr[xhrSymbols.properties];
-    const flag = xhr[xhrSymbols.flag];
-
-    properties.readyState = XMLHttpRequest.DONE;
-    properties.send = false;
-
-    setResponseToNetworkError(xhr);
-
-    if (flag.synchronous) {
-      throw exception;
-    }
-
-    xhr.dispatchEvent(new Event("readystatechange"));
-
-    if (!properties.uploadComplete) {
-      properties.uploadComplete = true;
-
-      if (properties.uploadListener) {
-        xhr.upload.dispatchEvent(new ProgressEvent(event));
-        xhr.upload.dispatchEvent(new ProgressEvent("loadend"));
-      }
-    }
-
-    xhr.dispatchEvent(new ProgressEvent(event));
-    xhr.dispatchEvent(new ProgressEvent("loadend"));
-  }
-
-  function setResponseToNetworkError(xhr) {
-    const properties = xhr[xhrSymbols.properties];
-    properties.responseCache = properties.responseTextCache = properties.responseXMLCache = null;
-    properties.responseHeaders = {};
-    properties.status = 0;
-    properties.statusText = "";
-  }
-
   return XMLHttpRequest;
 };
+
+function toByteString(value) {
+  value = String(value);
+  if (!/^[\0-\xFF]*$/.test(value)) {
+    throw new TypeError("invalid ByteString");
+  }
+  return value;
+}
+
+function finalMIMEType(xhr) {
+  const flag = xhr[xhrSymbols.flag];
+  return parseContentType(flag.overrideMIMEType || getResponseHeader(xhr, "content-type"));
+}
+
+function finalCharset(xhr) {
+  const flag = xhr[xhrSymbols.flag];
+  if (flag.overrideCharset) {
+    return flag.overrideCharset;
+  }
+  const parsedContentType = parseContentType(getResponseHeader(xhr, "content-type"));
+  if (parsedContentType) {
+    return whatwgEncoding.labelToName(parsedContentType.get("charset"));
+  }
+  return null;
+}
+
+function getResponseHeader(xhr, lcHeader) {
+  const properties = xhr[xhrSymbols.properties];
+  const keys = Object.keys(properties.responseHeaders);
+  let n = keys.length;
+  while (n--) {
+    const key = keys[n];
+    if (key.toLowerCase() === lcHeader) {
+      return properties.responseHeaders[key];
+    }
+  }
+  return null;
+}
+
+function normalizeHeaderValue(value) {
+  return value.replace(/^[\x09\x0A\x0D\x20]+/, "").replace(/[\x09\x0A\x0D\x20]+$/, "");
+}
 
 function coerceBodyArg(body) {
   // Implements the IDL conversion for `optional (Document or BodyInit)? body = null`
