+++ test/web-platform-tests/to-run.yaml
@@ -28,8 +28,6 @@ access-control-preflight-async-method-denied.htm: [fail, Unknown]
 access-control-preflight-async-not-supported.htm: [fail, Unknown]
 access-control-preflight-credential-async.htm: [fail, Unknown]
 access-control-preflight-credential-sync.htm: [fail, Unknown]
-access-control-preflight-headers-async.htm: [fail, Unknown]
-access-control-preflight-headers-sync.htm: [fail, Unknown]
 access-control-preflight-request-header-lowercase.htm: [fail, Unknown]
 access-control-preflight-request-header-sorted.htm: [fail, Unknown]
 access-control-preflight-sync-header-denied.htm: [fail, Unknown]
